move mergedTemplate to utils
